[serial] add separate serial port types
USB and TTL serial have subtle differences so we should have a way to
support this. This commit adds a different port constructor for USB
and TTL serial ports. The current difference is just in the handling
of DTR and RTS (which are not used on the PTI USB harness).