[feature] add liveness check
On startup, if the wrong port is provided or the target device
is not configured for RS232, the polling loop will continue to
run regardless of connection status. This uses a semaphore
to signal after a number of successful polling messages. Once
signalled, the acceptor's Start API will return true. Otherwise
the Start API will return false.